Create FUNDING.yml
My project is an open source collection of the best YouTube channels for nerds. My goal is to gather and organize the most interesting and educational content on topics such as science, technology, pop culture, history, art and more. I believe that my project can be a source of information and inspiration for people who like to learn new things and have fun with knowledge. I also want to encourage collaboration and participation from the open source community to suggest and evaluate the channels included in the collection.